+++ instrument-modules/user-modules/module-pradar-core/src/main/java/com/pamirs/pradar/utils/MonitorCollector.java
@@ -55,7 +55,6 @@ public class MonitorCollector {
     public static final String CPU_USAGE_KEY = "cpuUsage";
     public static final String IO_WAIT_KEY = "ioWait";
     public static final String NETWORK_BANDWIDTH_KEY = "networkBandwidth";
-    public static final String NETWORK_BANDWIDTH_RATE_KEY = "networkBandwidthRate";
     private final static Logger logger = LoggerFactory.getLogger(MonitorCollector.class.getName());
     private static final ThreadLocal<DecimalFormat> decimalFormatThreadLocal = new ThreadLocal<DecimalFormat>() {
         @Override
@@ -100,69 +99,6 @@ public class MonitorCollector {
         } else {
             executeResourcesInfoCollectingTaskOutsideContainer();
         }
-    }
-
-    private void executeResourcesInfoCollectingTaskInsideContainer() {
-        future = ExecutorServiceFactory.getFactory().scheduleAtFixedRate(new Runnable() {
-
-            private ContainerStatsInfoCollector collector;
-
-            @Override
-            public void run() {
-
-                if (!PradarSwitcher.isMonitorEnabled()) {
-                    return;
-                }
-                if (collector == null) {
-                    collector = new ContainerStatsInfoCollector();
-                }
-                try {
-                    long timeStamp = System.currentTimeMillis() / 1000;
-                    String appName = AppNameUtils.appName();
-                    StringBuilder stringBuilder = new StringBuilder();
-                    ContainerStatsInfoCollector.ContainerStatsInfo statsInfo = collector.getContainerStatsInfo();
-
-                    stringBuilder.append(appName).append("|").append(timeStamp).append("|");
-                    if (StringUtils.isNotBlank(Pradar.PRADAR_ENV_CODE)) {
-                        stringBuilder.append(
-                                StringUtils.isBlank(Pradar.PRADAR_TENANT_KEY) ? "" : Pradar.PRADAR_TENANT_KEY).append(
-                                "|")
-                            .append(StringUtils.isBlank(Pradar.PRADAR_ENV_CODE) ? "" : Pradar.PRADAR_ENV_CODE).append(
-                                "|")
-                            .append(StringUtils.isBlank(Pradar.PRADAR_USER_ID) ? "" : Pradar.PRADAR_USER_ID).append(
-                                "|");
-                    }
-                    stringBuilder.append(Pradar.AGENT_ID_NOT_CONTAIN_USER_INFO).append("|")
-                        .append(statsInfo.getCpuUsagePercent()).append("|")
-                        .append(statsInfo.getLatest1MinLoadAvg()).append("|")
-                        .append(statsInfo.getLatest5MinLoadAvg()).append("|")
-                        .append(statsInfo.getLatest15MinLoadAvg()).append("|")
-                        .append(statsInfo.getMemoryUsagePercent()).append("|")
-                        .append(statsInfo.getTotalMemory()).append("|")
-                        .append(statsInfo.getAvailableMemory()).append("|")
-                        .append(statsInfo.getCpuIoWaitUsagePercent()).append("|")
-                        .append(statsInfo.getNetworkUsage()).append('|')
-                        .append(statsInfo.getNetworkSpeed()).append('|')
-                        .append(statsInfo.getCoresNum()).append("|")
-                        .append(statsInfo.getTotalDiskSpace()).append('|')
-                        .append(statsInfo.getUsableDiskSpace()).append('|')
-                        .append(statsInfo.getDiskReadBytes()).append('|')
-                        .append(statsInfo.getDiskWriteBytes()).append('|');
-                    if (StringUtils.isNotBlank(Pradar.PRADAR_ENV_CODE)) {
-                        stringBuilder.append(1).append('|');
-                    }
-
-                    stringBuilder.append(Pradar.PRADAR_MONITOR_LOG_VERSION)
-                        .append(PradarCoreUtils.NEWLINE);
-                    Pradar.commitMonitorLog(stringBuilder.toString());
-                } catch (Throwable e) {
-                    if (printLogCount > 0) {
-                        printLogCount--;
-                        logger.error("write server monitor error!", e);
-                    }
-                }
-            }
-        }, 5, 1, TimeUnit.SECONDS);
     }
 
     private void executeResourcesInfoCollectingTaskOutsideContainer() {
@@ -234,6 +170,11 @@ public class MonitorCollector {
             }
         }, 5, 1, TimeUnit.SECONDS);
     }
+
+    private void executeResourcesInfoCollectingTaskInsideContainer() {
+        future = ExecutorServiceFactory.getFactory().scheduleAtFixedRate(new InsideContainerResourcesInfoCollector(), 5, 1, TimeUnit.SECONDS);
+    }
+
 
     public long[] getFileSystem() {
         try {
